Move stuff from Vec<AzleProgram> to TsAst
