+++ src/compiler/typescript_to_rust/azle_generate/src/generators/cross_canister_call_functions/mod.rs
@@ -1,15 +1,11 @@
-use crate::{
-    cdk_act::{SystemStructureType, ToActDataType, ToTokenStream},
-    ts_ast::{
-        azle_type::AzleType,
-        program::{azle_program::AzleProgramVecHelperMethods, AzleProgram},
-        AzleTypeAliasDecl, GetName,
-    },
-};
-use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 use swc_common::SourceMap;
 use swc_ecma_ast::{Expr, TsFnParam, TsMethodSignature, TsType, TsTypeElement, TsTypeLit};
+
+use crate::{
+    cdk_act::{ToActDataType, ToTokenStream},
+    ts_ast::{azle_type::AzleType, AzleTypeAliasDecl, GetName},
+};
 
 #[derive(Clone)]
 pub struct CrossCanisterCallFunctionsInfo {
@@ -42,81 +38,7 @@ struct CrossCanisterCallFunctionNames {
     notify_with_payment128_function_name: String,
 }
 
-pub trait CrossCanisterHelperMethods {
-    fn generate_cross_canister_call_functions(&self) -> TokenStream;
-    fn generate_cross_canister_call_functions_infos(&self) -> Vec<CrossCanisterCallFunctionsInfo>;
-}
-
-impl CrossCanisterHelperMethods for Vec<AzleProgram> {
-    fn generate_cross_canister_call_functions(&self) -> TokenStream {
-        let cross_canister_call_functions_infos =
-            self.generate_cross_canister_call_functions_infos();
-
-        let call_functions: Vec<proc_macro2::TokenStream> = cross_canister_call_functions_infos
-            .iter()
-            .map(|cross_canister_call_functions_info| {
-                cross_canister_call_functions_info.call.rust.clone()
-            })
-            .collect();
-
-        let call_with_payment_functions: Vec<proc_macro2::TokenStream> =
-            cross_canister_call_functions_infos
-                .iter()
-                .map(|cross_canister_call_functions_info| {
-                    cross_canister_call_functions_info
-                        .call_with_payment
-                        .rust
-                        .clone()
-                })
-                .collect();
-
-        let call_with_payment128_functions: Vec<proc_macro2::TokenStream> =
-            cross_canister_call_functions_infos
-                .iter()
-                .map(|cross_canister_call_functions_info| {
-                    cross_canister_call_functions_info
-                        .call_with_payment128
-                        .rust
-                        .clone()
-                })
-                .collect();
-
-        let notify_functions: Vec<proc_macro2::TokenStream> = cross_canister_call_functions_infos
-            .iter()
-            .map(|cross_canister_call_functions_info| {
-                cross_canister_call_functions_info.notify.rust.clone()
-            })
-            .collect();
-
-        let notify_with_payment128_functions: Vec<proc_macro2::TokenStream> =
-            cross_canister_call_functions_infos
-                .iter()
-                .map(|cross_canister_call_functions_info| {
-                    cross_canister_call_functions_info
-                        .notify_with_payment128
-                        .rust
-                        .clone()
-                })
-                .collect();
-
-        quote! {
-            #(#call_functions)*
-            #(#call_with_payment_functions)*
-            #(#call_with_payment128_functions)*
-            #(#notify_functions)*
-            #(#notify_with_payment128_functions)*
-        }
-    }
-
-    fn generate_cross_canister_call_functions_infos(&self) -> Vec<CrossCanisterCallFunctionsInfo> {
-        let canister_type_alias_decls = self
-            .get_azle_type_alias_decls_for_system_structure_type(&SystemStructureType::Canister);
-
-        canister_type_alias_decls.generate_cross_canister_call_functions_infos()
-    }
-}
-
-trait GenerateCrossCanisterCallFunctionsInfos {
+pub trait GenerateCrossCanisterCallFunctionsInfos {
     fn generate_cross_canister_call_functions_infos(&self) -> Vec<CrossCanisterCallFunctionsInfo>;
 }
