+++ src/compiler/typescript_to_rust/azle_generate/src/generators/mod.rs
@@ -1,4 +1,3 @@
-pub mod async_result_handler;
 pub mod canister_methods;
 pub mod cross_canister_call_functions;
 pub mod errors;
