+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/program/mod.rs
@@ -1,5 +1,2 @@
-mod errors;
-mod system_canister_method_builder;
-
 pub mod azle_program;
 pub use azle_program::AzleProgram;
