+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/cross_canister_calls.rs
@@ -1,0 +1,83 @@
+use proc_macro2::TokenStream;
+use quote::quote;
+
+use super::TsAst;
+use crate::{
+    cdk_act::SystemStructureType,
+    generators::cross_canister_call_functions::{
+        CrossCanisterCallFunctionsInfo, GenerateCrossCanisterCallFunctionsInfos,
+    },
+    ts_ast::program::azle_program::AzleProgramVecHelperMethods,
+};
+
+impl TsAst {
+    pub fn generate_cross_canister_call_functions(&self) -> TokenStream {
+        let cross_canister_call_functions_infos =
+            self.generate_cross_canister_call_functions_infos();
+
+        let call_functions: Vec<proc_macro2::TokenStream> = cross_canister_call_functions_infos
+            .iter()
+            .map(|cross_canister_call_functions_info| {
+                cross_canister_call_functions_info.call.rust.clone()
+            })
+            .collect();
+
+        let call_with_payment_functions: Vec<proc_macro2::TokenStream> =
+            cross_canister_call_functions_infos
+                .iter()
+                .map(|cross_canister_call_functions_info| {
+                    cross_canister_call_functions_info
+                        .call_with_payment
+                        .rust
+                        .clone()
+                })
+                .collect();
+
+        let call_with_payment128_functions: Vec<proc_macro2::TokenStream> =
+            cross_canister_call_functions_infos
+                .iter()
+                .map(|cross_canister_call_functions_info| {
+                    cross_canister_call_functions_info
+                        .call_with_payment128
+                        .rust
+                        .clone()
+                })
+                .collect();
+
+        let notify_functions: Vec<proc_macro2::TokenStream> = cross_canister_call_functions_infos
+            .iter()
+            .map(|cross_canister_call_functions_info| {
+                cross_canister_call_functions_info.notify.rust.clone()
+            })
+            .collect();
+
+        let notify_with_payment128_functions: Vec<proc_macro2::TokenStream> =
+            cross_canister_call_functions_infos
+                .iter()
+                .map(|cross_canister_call_functions_info| {
+                    cross_canister_call_functions_info
+                        .notify_with_payment128
+                        .rust
+                        .clone()
+                })
+                .collect();
+
+        quote! {
+            #(#call_functions)*
+            #(#call_with_payment_functions)*
+            #(#call_with_payment128_functions)*
+            #(#notify_functions)*
+            #(#notify_with_payment128_functions)*
+        }
+    }
+
+    pub fn generate_cross_canister_call_functions_infos(
+        &self,
+    ) -> Vec<CrossCanisterCallFunctionsInfo> {
+        let canister_type_alias_decls = self
+            .azle_programs
+            .get_azle_type_alias_decls_for_system_structure_type(&SystemStructureType::Canister);
+
+        canister_type_alias_decls.generate_cross_canister_call_functions_infos()
+    }
+}
