+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/system_canister_method_builder/heartbeat.rs
@@ -1,13 +1,13 @@
 use crate::{
     cdk_act::{nodes::ActHeartbeatMethod, CanisterMethodType},
     generators::canister_methods::method_body,
-    ts_ast::program::{azle_program::AzleProgramVecHelperMethods, errors, AzleProgram},
+    ts_ast::{program::azle_program::AzleProgramVecHelperMethods, ts_ast::errors, TsAst},
 };
 
-pub fn build_canister_method_system_heartbeat(
-    programs: &Vec<AzleProgram>,
-) -> Option<ActHeartbeatMethod> {
-    let heartbeat_fn_decls = programs.get_azle_fn_decls_of_type(&CanisterMethodType::Heartbeat);
+pub fn build_canister_method_system_heartbeat(ts_ast: &TsAst) -> Option<ActHeartbeatMethod> {
+    let heartbeat_fn_decls = ts_ast
+        .azle_programs
+        .get_azle_fn_decls_of_type(&CanisterMethodType::Heartbeat);
 
     if heartbeat_fn_decls.len() > 1 {
         let error_message = errors::create_duplicate_method_types_error_message(
