+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/system_canister_method_builder/inspect_message.rs
@@ -1,14 +1,15 @@
 use crate::{
     cdk_act::{nodes::ActInspectMessageMethod, CanisterMethodType},
     generators::canister_methods::method_body,
-    ts_ast::program::{azle_program::AzleProgramVecHelperMethods, errors, AzleProgram},
+    ts_ast::{program::azle_program::AzleProgramVecHelperMethods, ts_ast::errors, TsAst},
 };
 
 pub fn build_canister_method_system_inspect_message(
-    programs: &Vec<AzleProgram>,
+    ts_ast: &TsAst,
 ) -> Option<ActInspectMessageMethod> {
-    let inspect_message_fn_decls =
-        programs.get_azle_fn_decls_of_type(&CanisterMethodType::InspectMessage);
+    let inspect_message_fn_decls = ts_ast
+        .azle_programs
+        .get_azle_fn_decls_of_type(&CanisterMethodType::InspectMessage);
 
     if inspect_message_fn_decls.len() > 1 {
         let error_message = errors::create_duplicate_method_types_error_message(
