+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/system_canister_method_builder/post_upgrade.rs
@@ -1,21 +1,17 @@
 use quote::quote;
 
 use crate::{
-    cdk_act::{
-        generators::ic_object::IcObjectHelperMethods, nodes::ActPostUpgradeMethod,
-        traits::CanisterMethodBuilder, CanisterMethodType,
-    },
+    cdk_act::{nodes::ActPostUpgradeMethod, traits::CanisterMethodBuilder, CanisterMethodType},
     generators::canister_methods::method_body,
-    ts_ast::program::{azle_program::AzleProgramVecHelperMethods, errors, AzleProgram},
+    ts_ast::{program::azle_program::AzleProgramVecHelperMethods, ts_ast::errors, TsAst},
 };
 
-pub fn build_canister_method_system_post_upgrade(
-    programs: &Vec<AzleProgram>,
-) -> ActPostUpgradeMethod {
-    let ic_object = programs.generate_ic_object();
+pub fn build_canister_method_system_post_upgrade(ts_ast: &TsAst) -> ActPostUpgradeMethod {
+    let ic_object = ts_ast.generate_ic_object();
 
-    let post_upgrade_fn_decls =
-        programs.get_azle_fn_decls_of_type(&CanisterMethodType::PostUpgrade);
+    let post_upgrade_fn_decls = ts_ast
+        .azle_programs
+        .get_azle_fn_decls_of_type(&CanisterMethodType::PostUpgrade);
 
     if post_upgrade_fn_decls.len() > 1 {
         let error_message = errors::create_duplicate_method_types_error_message(
